Further implementation of the filesystem and VFS

+++ libc/include/stdint.h
@@ -9,17 +9,17 @@
 #define SIZE_MAX 	4294967295 	// (2**32)-1
 
 /* signed integral types */
-typedef char  int8_t;
-typedef short int16_t;
-typedef int   int32_t;
-typedef long  int64_t;
-typedef long  intmax_t;
+typedef char        int8_t;
+typedef short       int16_t;
+typedef int         int32_t;
+typedef long long   int64_t;
+typedef long long   intmax_t;
 
 /* unsigned integral types */
-typedef unsigned char  uint8_t;
-typedef unsigned short uint16_t;
-typedef unsigned int   uint32_t;
-typedef unsigned long  uint64_t;
-typedef unsigned long  uintmax_t;
+typedef unsigned char       uint8_t;
+typedef unsigned short      uint16_t;
+typedef unsigned int        uint32_t;
+typedef unsigned long long  uint64_t;
+typedef unsigned long       uintmax_t;
 
 #endif
